requirements.txt
bidict==0.23.1
blinker==1.7.0
click==8.1.7
Flask==3.0.3
Flask-Cors==4.0.0
Flask-SocketIO==5.3.6
h11==0.14.0
importlib_metadata==7.1.0
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.4.0
MarkupSafe==2.1.5
numpy==1.26.4
PyJWT==2.8.0
python-engineio==4.9.0
python-socketio==5.11.2
scikit-learn==1.4.2
scipy==1.13.0
simple-websocket==1.0.0
threadpoolctl==3.4.0
tinydb==4.8.0
Werkzeug==3.0.2
wsproto==1.2.0
zipp==3.18.1
